+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## Release Version
 
-- [minor] Updated MoEngage-iOS-SDK to 10.11.0
+- [major] Added Personalize Experience APIs support